Pool.c is splitted to three files
